# asklepios-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Asklepios ERP API Documentation",
        "description": "Documentation interactive de l'API Asklepios pour la gestion hospitalière. Permet aux développeurs front-end de tester les requêtes en temps réel.",
        "contact": {
            "name": "Support Technique Asklepios",
            "email": "admin@asklepios.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur API de Développement"
        }
    ],
    "paths": {
...
        "/api/admin/articles": {
            "get": {
                "tags": [
                    "Articles (Admin)"
                ],
                "summary": "Lister les articles (Paginés)",
                "description": "Lister et rechercher des articles (Paginés)",
                "operationId": "getAdminArticles",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Nom ou Code-barres",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre de résultats par page (défaut: 10)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de la page à récupérer (défaut: 1)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste paginée récupérée avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Articles (Admin)"
                ],
                "summary": "Créer un article",
                "description": "Créer un nouvel article",
                "operationId": "storeAdminArticle",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "category_id",
                                    "name",
                                    "track_batches"
                                ],
                                "properties": {
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "barcode": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "global_min_qty": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "track_batches": {
                                        "description": "Vrai si l'article gère des lots classiques",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Article créé avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/articles/all": {
            "get": {
                "tags": [
                    "Articles (Admin)"
                ],
                "summary": "Lister tous les articles (Sans pagination)",
                "description": "Lister tous les articles (Sans pagination, pour les select)",
                "operationId": "getAllAdminArticles",
                "responses": {
                    "200": {
                        "description": "Liste complète récupérée avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/articles/{id}": {
            "post": {
                "tags": [
                    "Articles (Admin)"
                ],
                "summary": "Modifier un article",
                "description": "Modifier un article",
                "operationId": "updateAdminArticle",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article modifié avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Articles (Admin)"
                ],
                "summary": "Supprimer un article",
                "description": "Supprimer un article",
                "operationId": "deleteAdminArticle",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article supprimé avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Saisissez le jeton (Token) obtenu lors de la connexion pour tester les routes protégées.",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}